chore: move module related code from `fm` to `noirc-frontend`
