+++ compiler/fm/Cargo.toml
@@ -12,5 +12,5 @@ codespan-reporting.workspace = true
 serde.workspace = true
 
 [dev-dependencies]
-tempfile = "3.2.0"
+tempfile.workspace = true
 iter-extended.workspace = true
