+++ compiler/noirc_frontend/Cargo.toml
@@ -26,3 +26,4 @@ regex = "1.9.1"
 [dev-dependencies]
 strum = "0.24"
 strum_macros = "0.24"
+tempfile.workspace = true
