+++ tooling/nargo/Cargo.toml
@@ -29,4 +29,4 @@ rayon = "1.8.0"
 [dev-dependencies]
 # TODO: This dependency is used to generate unit tests for `get_all_paths_in_dir`
 # TODO: once that method is moved to nargo_cli, we can move this dependency to nargo_cli
-tempfile = "3.2.0"
+tempfile.workspace = true
