+++ tooling/nargo_cli/Cargo.toml
@@ -57,7 +57,7 @@ bb_abstraction_leaks.workspace = true
 tokio-util = { version = "0.7.8", features = ["compat"] }
 
 [dev-dependencies]
-tempfile = "3.6.0"
+tempfile.workspace = true
 dirs.workspace = true
 assert_cmd = "2.0.8"
 assert_fs = "1.0.10"
